requirements.txt
# IBM watsonx
ibm-watsonx-ai>=1.0.0
ibm-cloud-sdk-core>=3.18.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Enterprise Tool Integrations
slack-sdk>=3.23.0
PyGithub>=2.1.0
atlassian-python-api>=3.41.0
pysnow>=0.7.17  # ServiceNow

# HTTP & Async
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# AI & LLM
langchain>=0.1.0
langchain-ibm>=0.1.0
openai>=1.0.0  # Fallback option

# Data Processing
pandas>=2.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# Logging & Monitoring
structlog>=23.2.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.0
tenacity>=8.2.0  # Retry logic
